chore: small refactor
