+++ apis/extractor/src/config.ts
@@ -434,5 +434,20 @@ export const EXTRACTOR_CONFIG = {
   // },
 }
 
-export type CONFIGURED_CHAIN_IDS = keyof typeof EXTRACTOR_CONFIG
+export const PORT = process.env['PORT'] || 80
 
+export const SENTRY_DSN = process.env['SENTRY_DSN'] as string
+if (!SENTRY_DSN) {
+  throw new Error('SENTRY_DSN is not set')
+}
+export const SENTRY_ENVIRONMENT = process.env['SENTRY_ENVIRONMENT'] as string
+if (!SENTRY_ENVIRONMENT) {
+  throw new Error('SENTRY_ENVIRONMENT is not set')
+}
+
+export const CHAIN_ID = Number(
+  process.env['CHAIN_ID'],
+) as keyof typeof EXTRACTOR_CONFIG
+if (!CHAIN_ID) {
+  throw new Error('CHAIN_ID is not set')
+}
